Import SHACL from `app-lblod-harvester`.
These are the SHACL shapes for the application profile for OSLO
Besluitvorming, as of its state in:

lblod/app-lblod-harvester@1b369942f5d26be627e3dc56af5fb829afd4d90b

We will use ELI{,-DL}, so these will not be used. Nevertheless, the file
is imported. One reason is to elucidate the history, but the more
important one is that it makes it easier to transplant all commits
previously made to the 'wrong' repository:

<https://github.com/lblod/app-eli-dl-harvester/pull/1>

For completeness and posterity, I will also copy the description of the
pull request:

> These SHACL shapes are based on the existing SHACL shapes for OSLO
Besluitvorming, with a mapping applied for ELI-DL as described in
[GitBook](https://app.gitbook.com/o/-MP9Yduzf5xu7wIebqPG/s/grRYjcouX0uZY2P6PYBl/data-analysis-research/map-oslo-to-eli-dl).

> There are some classes and properties that are present in the OSLO
Besluitvorming SHACL shapes, but that have no mapping yet. Where I
could, I have added these mappings myself in commit
[3d26e3f](https://github.com/lblod/app-eli-dl-harvester/pull/1/commits/3d26e3ff70318137430dc6729c846aba299e76e0);
perhaps extra attention is merited there. When I was unsure, I have
added a \# TODO comment. What shall we do with these? Perhaps they are
unused in our application profile anyway.

> There are also some mappings that do have a mapping, but that were not
present in OSLO Besluitvorming SHACL shapes. I have indicated them in a
comment at the top.

> Eventually, these shapes should go to the app repository that
@MartijnBogaert created at https://github.com/lblod/app-decide/pull/3.
They initially went into this repository because Nordine suggested I'd
work in a fork --- and for now, I've kept them here, because the git
history might be instructive for reviewing. Perhaps transplant these
commits when done.